fix: 피드백 사항 반영

+++ src/main/java/kr/co/suitcarrier/web/dto/LikeRequestDto.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 @Getter
 public class LikeRequestDto {
 
-    private Integer user;
-    private Integer post;
+    private Long user;
+    private Long post;
 
 }
